Remove FSharpProjectSystemPackage
