+++ src/Microsoft.VisualStudio.ProjectSystem.FSharp.VS/Microsoft.VisualStudio.ProjectSystem.FSharp.VS.csproj
@@ -4,7 +4,7 @@
   <PropertyGroup>
     <RootNamespace>Microsoft.VisualStudio</RootNamespace>
     <!-- VSIX -->
-    <GeneratePkgDefFile>true</GeneratePkgDefFile>
+    <GeneratePkgDefFile>false</GeneratePkgDefFile>
     <CreateVsixContainer>false</CreateVsixContainer>
   </PropertyGroup>
   <ItemGroup>
@@ -14,11 +14,5 @@
   <ItemGroup>
     <InternalsVisibleTo Include="Microsoft.VisualStudio.ProjectSystem.FSharp.VS.UnitTests" />
     <InternalsVisibleTo Include="DynamicProxyGenAssembly2" Key="$(MoqPublicKey)" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="xlf\*" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="xlf\" />
   </ItemGroup>
 </Project>
