+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS/PublicAPI.Unshipped.txt
@@ -1,3 +1,6 @@
+Microsoft.VisualStudio.ProjectSystem.VS.FSharp.FSharpProjectSelector
+Microsoft.VisualStudio.ProjectSystem.VS.FSharp.FSharpProjectSelector.FSharpProjectSelector() -> void
+Microsoft.VisualStudio.ProjectSystem.VS.FSharp.FSharpProjectSelector.GetProjectFactoryGuid(System.Guid guidProjectType, string pszFilename, out System.Guid guidProjectFactory) -> void
 static Microsoft.VisualStudio.ProjectSystem.VS.PropertyPages.PropertyPageResources.CancelBtnHelpText.get -> string
 static Microsoft.VisualStudio.ProjectSystem.VS.PropertyPages.PropertyPageResources.DeleteBtnHelpText.get -> string
 static Microsoft.VisualStudio.ProjectSystem.VS.PropertyPages.PropertyPageResources.NewBtnHelpText.get -> string
