+++ src/ProjectSystemSetup/ProjectSelectors.pkgdef
@@ -5,4 +5,4 @@
 [$RootKey$\Projects\{F2A71F9B-5D33-465A-A702-920D77279786}]
 "ProjectSelector"="{E720DAD0-1854-47FC-93AF-E719B54B84E6}"
 [$RootKey$\ProjectSelectors\{E720DAD0-1854-47FC-93AF-E719B54B84E6}]
-"Package"="{A724C878-E8FD-4FEB-B537-60BABA7EDA83}"
+"Package"="{A4F9D880-9492-4072-8BF3-2B5EEEDC9E68}"
